pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "apk2img-ml"
version = "0.1.0"
description = "APK token extraction, Doc2Vec workflows, and image conversion tools."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "gensim>=4.3",
  "numpy>=1.24",
  "pillow>=10.0",
  "tqdm>=4.66",
  "androguard>=4.1",
]

[project.optional-dependencies]
cnn = [
  "torch>=2.0",
  "torchvision>=0.15",
]
baksmali = []
sdhash = []

[project.scripts]
apk2img-ml = "apk2img_ml.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["apk2img_ml*"]
